refactor: 타입 파일 경로 이동
박경은 님의 조언에 따라 타입 파일들을 각각의 모듈  폴더로 이동

closes #71
